move custom gettext.lua to /l10n
