+++ _meta.lua
@@ -1,4 +1,4 @@
-local _ = require("resources.gettext")
+local _ = require("l10n.gettext")
 return {
     -- we have to lie here and claim to be coverbrowser now that there are explicit calls from core koreader
     -- name = "projecttitle",
