+++ altbookstatuswidget.lua
@@ -21,7 +21,7 @@ local VerticalGroup = require("ui/widget/verticalgroup")
 local VerticalSpan = require("ui/widget/verticalspan")
 local logger = require("logger")
 local util = require("util")
-local _ = require("resources.gettext")
+local _ = require("l10n.gettext")
 local Screen = Device.screen
 local T = require("ffi/util").template
 local BookInfoManager = require("bookinfomanager")
