+++ covermenu.lua
@@ -24,7 +24,7 @@ local TextWidget = require("ui/widget/textwidget")
 local TitleBar = require("titlebar")
 local FrameContainer = require("ui/widget/container/framecontainer")
 local FileManagerMenu = require("apps/filemanager/filemanagermenu")
-local _ = require("resources.gettext")
+local _ = require("l10n.gettext")
 local Device = require("device")
 local filemanagerutil = require("apps/filemanager/filemanagerutil")
 local logger = require("logger")
