+++ listmenu.lua
@@ -30,7 +30,7 @@ local logger = require("logger")
 local lfs = require("libs/libkoreader-lfs")
 local util = require("util")
 local time = require("ui/time")
-local _ = require("resources.gettext")
+local _ = require("l10n.gettext")
 local Screen = Device.screen
 local T = require("ffi/util").template
 local getMenuText = require("ui/widget/menu").getMenuText
