+++ mosaicmenu.lua
@@ -31,7 +31,7 @@ local lfs = require("libs/libkoreader-lfs")
 local logger = require("logger")
 local util = require("util")
 local time = require("ui/time")
-local _ = require("resources.gettext")
+local _ = require("l10n.gettext")
 local Screen = Device.screen
 local T = require("ffi/util").template
 local getMenuText = require("ui/widget/menu").getMenuText
